+++ etc/uikit-playground/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @milaboratories/uikit-playground
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [add109d]
+  - @milaboratories/uikit@1.2.7
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [d9f5f2b]
+  - @milaboratories/uikit@1.2.6
+
 ## 0.0.3
 
 ### Patch Changes